README.md
The "Wine Quality Prediction" project employs machine learning techniques to develop a 
predictive model for assessing and predicting the quality of wines. This project is motivated by 
the need within the wine industry to consistently produce high-quality wines that align with 
consumer preferences. By leveraging historical data on various wine attributes, the project aims 
to create a model capable of providing winemakers with valuable insights into the quality of 
their products.